+++ prevedere/__init__.py
@@ -1,1 +1,1 @@
-from .prevedere import *
+from .api import Api